Enable fc node to use different drivers
