+++ packages/flight_controller_driver/setup.py
@@ -5,7 +5,7 @@ from catkin_pkg.python_setup import generate_distutils_setup
 
 # fetch values from package.xml
 setup_args = generate_distutils_setup(
-    packages=['h2r_multiwii'],
+    packages=['fc_driver'],
     package_dir={'': 'include'},
 )
